+++ aro-service/aro-engine/src/main/java/com/altvil/aro/service/planing/WirecenterNetworkPlan.java
@@ -7,7 +7,7 @@ import com.altvil.aro.model.NetworkNode;
 
 public interface WirecenterNetworkPlan {
 
-	int getPlanId();
+	long getPlanId();
 
 	Collection<NetworkNode> getNetworkNodes();
